# openapi.yaml
openapi: 3.1.0
info:
  title: NextWave Voice Logistics Demo API
  version: 0.1.0
  summary: API para una demo de coordinación logística terrestre mediante agentes de voz.
  description: |
    Contrato HTTP del monolito Express para la demo NextWave Hackathon 2026.

    La API mantiene las decisiones verificables en el backend: mandato, ofertas,
    selección, commitments y estados. OpenAI Realtime controla la conversación y
    utiliza estos endpoints como tools. Twilio entrega telefonía y webhooks.

    Alcance deliberado de demo: SQLite, cola en memoria de Node.js, una sola
    instancia, sin autenticación real y sin garantías de recuperación después de
    reiniciar el proceso.
  contact:
    name: NextWave demo team
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: Servidor local de la demo
...
paths:
...
  /operations/{operationId}/mandate:
    parameters:
      - $ref: "#/components/parameters/OperationId"
    get:
      tags: [Mandates]
      operationId: getActiveMandate
      summary: Consultar el mandato vigente
      description: Tool de solo lectura disponible para el Logistics Agent.
      responses:
        "200":
          description: Mandato activo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mandate"
        "404":
          $ref: "#/components/responses/NotFound"
  /operations/{operationId}/mandates/versions:
    parameters:
      - $ref: "#/components/parameters/OperationId"
    post:
      tags: [Mandates]
      operationId: createMandateVersion
      summary: Crear una nueva versión inmutable del mandato
      description: El Logistics Agent nunca recibe esta tool.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMandateVersionRequest"
      responses:
        "201":
          description: Nueva versión activa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mandate"
        "409":
          $ref: "#/components/responses/Conflict"
  /carriers:
    get:
      tags: [Carriers]
      operationId: listCarriers
      summary: Listar carriers disponibles para la campaña
      responses:
        "200":
          description: Carriers de la demo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Carrier"
    post:
      tags: [Carriers]
      operationId: createCarrier
      summary: Crear un carrier ficticio o de prueba
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarrierRequest"
      responses:
        "201":
          description: Carrier creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Carrier"
...
components:
  parameters:
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
      example: op_01JABC123
...
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: La acción no es compatible con el estado actual
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_STATE_TRANSITION
        message:
          type: string
          example: La operación no puede pasar de CREATED a IN_TRANSIT.
        details:
          type: object
          additionalProperties: true
...
    CreateMandateRequest:
      type: object
      required: [maxTotalPrice, currency, pickupDate]
      properties:
        maxTotalPrice:
          type: number
          exclusiveMinimum: 0
          example: 9000
        currency:
          type: string
          example: MXN
        pickupDate:
          type: string
          format: date
          example: 2026-09-03
        notes:
          type: string
    CreateMandateVersionRequest:
      $ref: "#/components/schemas/CreateMandateRequest"
    Mandate:
      allOf:
        - $ref: "#/components/schemas/CreateMandateRequest"
        - type: object
          required: [id, operationId, version, status, createdAt]
          properties:
            id:
              type: string
              example: man_01JABC123
            operationId:
              type: string
              example: op_01JABC123
            version:
              type: integer
              minimum: 1
            status:
              type: string
              enum: [ACTIVE, SUPERSEDED]
            createdAt:
              type: string
              format: date-time
    CreateCarrierRequest:
      type: object
      required: [name, dispatcherName, phone]
      properties:
        name:
          type: string
          example: Transportes Atlas
        dispatcherName:
          type: string
          example: Laura
        phone:
          type: string
          example: "+525555555501"
        email:
          type: string
          format: email
          example: operaciones@atlas.example
        score:
          type: number
          minimum: 0
          maximum: 100
          default: 80
    Carrier:
      allOf:
        - $ref: "#/components/schemas/CreateCarrierRequest"
        - type: object
          required: [id, active]
          properties:
            id:
              type: string
              example: car_atlas
            active:
              type: boolean
              default: true
...